♻️ refactor: teamEdit 모달 디자인 개선
