+++ app/_components/modal-contents/TeamEdit.tsx
@@ -5,6 +5,8 @@ import { useMutation, useQuery, useQueryClient } from "@tanstack/react-query";
 import Image from "next/image";
 import { useRouter } from "next/navigation";
 import { useCallback } from "react";
+import CloseIcon from "@/public/icons/ic_close";
+import Icon from "@/app/_icons";
 
 import API from "@/app/_api";
 
@@ -76,14 +78,20 @@ export default function TeamEdit({ close, id, initialTeamName }: TeamEditProps):
 
 	return (
 		<ModalWrapper close={close}>
-			<section className="flex size-full flex-col items-center justify-center">
-				<div className="pb-20">
-					<h1 className="text-[40px] font-medium leading-[48px] text-white">팀 수정하기</h1>
+			<section className="flex size-full flex-col items-center justify-center px-[20px] pb-[15px]">
+				<div className="flex w-full justify-end">
+					<button onClick={close} type="button" aria-label="Close modal">
+						<CloseIcon width={24} height={24} />
+					</button>
+				</div>
+
+				<div className="pb-7">
+					<h1 className="text-[18px] font-medium text-white">팀 수정하기</h1>
 				</div>
 
 				<div className="w-full">
 					<Form onSubmit={handleTeamManagement}>
-						<div className="mx-auto flex size-full max-w-[340px] flex-col tablet:max-w-[460px]">
+						<div className="flex flex-col">
 							<div className="w-fit">
 								<label htmlFor="team-name" className="w-full text-start text-text-primary">
 									팀 프로필
@@ -92,7 +100,7 @@ export default function TeamEdit({ close, id, initialTeamName }: TeamEditProps):
 								<Form.ImageInput id="profileImage" tests={[{ type: "file_size", data: 1048576, error: "이미지 파일 크기는 1MB 이하여야 합니다" }]}>
 									{(file) =>
 										file || teamInfo?.image ? (
-											<div className="relative flex size-16 items-center justify-center rounded-[12px] border-2 border-border-primary/10">
+											<div className="relative flex size-16 cursor-pointer items-center justify-center rounded-[12px] border-2 border-border-primary/10">
 												<Image
 													src={(file as string) ?? teamInfo?.image ?? ""}
 													alt="Profile Preview"
@@ -100,16 +108,19 @@ export default function TeamEdit({ close, id, initialTeamName }: TeamEditProps):
 													className="rounded-[12px] object-cover object-center"
 												/>
 												<div className="relative size-full">
-													<Image src="/icons/editIcon.svg" alt="Profile Preview" width={20} height={20} className="absolute -bottom-2 -right-2" />
+													<div className="absolute -bottom-2 -right-2 flex h-[20px] w-[20px] items-center justify-center rounded-full border-2 border-[#1E293B] bg-[#334155]">
+														<Icon.Edit width={11} height={11} color="#64748B" />
+													</div>
 												</div>
 											</div>
 										) : (
-											<div className="relative flex size-16 items-center justify-center rounded-[12px] border-2 border-border-primary/10 bg-background-secondary">
-												<div className="relative size-5">
-													<Image src="/icons/imageIcon.svg" alt="Profile Preview" fill />
+											<div className="relative flex size-16 cursor-pointer items-center justify-center rounded-[12px] border-2 border-border-primary/10 bg-background-secondary">
+												<div className="\\ relative">
+													<Image src="/icons/emptyImage.svg" alt="Profile Preview" width={20} height={20} />
 												</div>
-
-												<Image src="/icons/editIcon.svg" alt="Profile Preview" width={20} height={20} className="absolute -bottom-2 -right-2" />
+												<div className="absolute -bottom-2 -right-2 flex h-[20px] w-[20px] items-center justify-center rounded-full border-2 border-[#1E293B] bg-[#334155] pl-[1px]">
+													<Icon.Edit width={11} height={11} color="#64748B" />
+												</div>
 											</div>
 										)
 									}
@@ -122,13 +133,16 @@ export default function TeamEdit({ close, id, initialTeamName }: TeamEditProps):
 							<label htmlFor="team-name" className="w-full pb-3 text-start text-text-primary">
 								팀 이름
 							</label>
-							<Form.Input
-								id="teamName"
-								type="text"
-								placeholder="팀 이름을 입력하세요"
-								init={initialTeamName}
-								tests={[{ type: "require", data: true, error: "팀 이름은 필수입니다" }]}
-							/>
+							<div className="">
+								<Form.Input
+									id="teamName"
+									type="text"
+									placeholder="팀 이름을 입력하세요"
+									init={initialTeamName}
+									tests={[{ type: "require", data: true, error: "팀 이름은 필수입니다" }]}
+								/>
+							</div>
+
 							<Form.Error htmlFor="team-name" />
 
 							<div className="pt-10" />
@@ -139,7 +153,7 @@ export default function TeamEdit({ close, id, initialTeamName }: TeamEditProps):
 				</div>
 
 				<div className="max-w-[340px] pt-6 tablet:max-w-[460px]">
-					<p className="text-md font-normal text-text-primary tablet:text-lg">팀 이름은 회사명이나 모임 이름 등으로 설정하면 좋아요.</p>
+					<p className="text-md font-normal text-text-primary tablet:text-[15px]">팀 이름은 회사명이나 모임 이름 등으로 설정하면 좋아요.</p>
 				</div>
 			</section>
 		</ModalWrapper>
